backend/requirements.txt
# ==================================
# CORE - FastAPI
# ==================================
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# ==================================
# BASE DE DATOS
# ==================================
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# ==================================
# VALIDACIÓN Y SCHEMAS
# ==================================
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.2.0

# ==================================
# GOOGLE VISION API
# ==================================
google-cloud-vision==3.9.0

# ==================================
# PROCESAMIENTO DE IMÁGENES
# ==================================
Pillow==11.1.0
pdf2image==1.17.0

# ==================================
# UTILIDADES
# ==================================
python-dotenv==1.0.1
python-dateutil==2.9.0

# ==================================
# HTTPX (para requests async)
# ==================================
httpx==0.28.1

# ==================================
# CORS Y SEGURIDAD
# ==================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ==================================
# TESTING (opcional)
# ==================================
pytest==8.3.4
pytest-asyncio==0.24.0